requirements.txt
matplotlib==3.7.1
Pillow==9.4.0
torch==1.12.1
torchsummary==1.5.1
torchvision==0.13.1
tqdm==4.65.0
pandas==2.0.0
